% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dobin-package.R
\docType{package}
\name{dobin-package}
\alias{dobin-package}
\title{dobin: Dimension Reduction for Outlier Detection}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A dimension reduction technique for outlier detection. DOBIN: a Distance based Outlier BasIs using Neighbours, constructs a set of basis vectors for outlier detection. This is not an outlier detection method; rather it is a pre-processing method for outlier detection. It brings outliers to the fore-front using fewer basis vectors (Kandanaarachchi, Hyndman 2020) <doi:10.1080/10618600.2020.1807353>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sevvandi.github.io/dobin/}
}

}
\author{
\strong{Maintainer}: Sevvandi Kandanaarachchi \email{sevvandik@gmail.com} (\href{https://orcid.org/0000-0002-0337-0395}{ORCID})

}
\keyword{internal}
